+++ lscolors/lscolors.go
@@ -16,8 +16,17 @@ import (
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
-// RESET_SEQ is ANSI reset sequence
-const RESET_SEQ = "\033[0m"
+const (
+	RESET  = "rs" // Reset
+	DIR    = "di" // Directory
+	LINK   = "ln" // Symbolic link
+	FIFO   = "pi" // Pipe
+	SOCK   = "so" // Socket
+	BLK    = "bd" // Block device driver
+	CHR    = "cd" // Character device driver
+	STICKY = "st" // Dir with the sticky bit set (+t) and not other-writable
+	EXEC   = "ex" // Executable files
+)
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
@@ -29,14 +38,23 @@ var initialized bool
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
+// DisableColors disables all colors in output
+var DisableColors = os.Getenv("NO_COLOR") != ""
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+
 // GetColor returns ANSI control sequence with color for given file
 func GetColor(file string) string {
 	if !initialized {
 		initialize()
 	}
 
-	if len(colorMap) == 0 {
+	if DisableColors || len(colorMap) == 0 {
 		return ""
+	}
+
+	if colorMap[file] != "" {
+		return "\033[" + colorMap[file] + "m"
 	}
 
 	for glob, color := range colorMap {
@@ -58,7 +76,7 @@ func Colorize(file string) string {
 		return file
 	}
 
-	return colorSeq + file + RESET_SEQ
+	return colorSeq + file + "\033[0m"
 }
 
 // Colorize return path with ANSI control sequences
@@ -70,7 +88,7 @@ func ColorizePath(fullPath string) string {
 		return fullPath
 	}
 
-	return colorSeq + fullPath + RESET_SEQ
+	return colorSeq + fullPath + "\033[0m"
 }
 
 // ////////////////////////////////////////////////////////////////////////////////// //
@@ -79,16 +97,20 @@ func ColorizePath(fullPath string) string {
 func initialize() {
 	initialized = true
 
+	if DisableColors {
+		return
+	}
+
 	lsColors := os.Getenv("LS_COLORS")
 
 	if lsColors == "" {
 		return
 	}
 
-	colorMap = make(map[string]string)
+	colorMap = map[string]string{RESET: "0"}
 
 	for _, key := range strings.Split(lsColors, ":") {
-		if !strings.HasPrefix(key, "*") || !strings.ContainsRune(key, '=') {
+		if !strings.ContainsRune(key, '=') || !strings.ContainsRune(key, ';') {
 			continue
 		}